+++ src/differential_privacy/aggregates.rs
@@ -4,7 +4,7 @@ use crate::{
     differential_privacy::private_query::PrivateQuery,
     differential_privacy::{private_query, DPRelation, Error, Result},
     expr::{aggregate, AggregateColumn, Expr},
-    protection::PEPRelation,
+    privacy_unit_tracking::PUPRelation,
     relation::{field::Field, Map, Reduce, Relation, Variant as _},
     DataType, Ready,
 };
@@ -30,6 +30,10 @@ impl Field {
 
 impl Relation {
     fn gaussian_mechanisms(self, epsilon: f64, delta: f64, bounds: Vec<(&str, f64)>) -> DPRelation {
+        if epsilon>1. {
+            // Cf. Theorem A.1. in (Dwork, Roth et al. 2014)
+            log::warn!("Warning, epsilon>1 the gaussian mechanism applied will not be exactly epsilon,delta-DP!")
+        }
         let number_of_agg = bounds.len() as f64;
         let noise_multipliers = bounds
             .into_iter()
@@ -57,7 +61,7 @@ impl Relation {
     }
 }
 
-impl PEPRelation {
+impl PUPRelation {
     /// Builds a DPRelation wrapping a Relation::Reduce
     /// whose `aggregates` are the noisy sums of each column in `sums`
     /// and the group by columns are defined by `group_by_names`
@@ -90,7 +94,7 @@ impl PEPRelation {
 
         // Clip the relation
         let clipped_relation = self.deref().clone().l2_clipped_sums(
-            self.protected_entity_id(),
+            self.privacy_unit(),
             group_by_names,
             input_values_bound.clone(),
         );
@@ -115,18 +119,15 @@ impl PEPRelation {
         let mut named_sums = vec![];
         let mut input_builder = Map::builder()
             .with((
-                self.protected_entity_id(),
+                self.privacy_unit(),
                 Expr::coalesce(
-                    Expr::cast_as_text(Expr::col(self.protected_entity_id())),
-                    Expr::val(self.protected_entity_null_id().to_string())
+                    Expr::cast_as_text(Expr::col(self.privacy_unit())),
+                    Expr::val(self.privacy_unit_default().to_string()),
                 ),
             ))
             .with((
-                self.protected_entity_weight(),
-                Expr::coalesce(
-                    Expr::col(self.protected_entity_weight()),
-                    Expr::val(0.)
-                )
+                self.privacy_unit_weight(),
+                Expr::coalesce(Expr::col(self.privacy_unit_weight()), Expr::val(0.)),
             ));
 
         let mut group_by_names = vec![];
@@ -186,9 +187,9 @@ impl PEPRelation {
         );
 
         let input: Relation = input_builder.input(self.deref().clone()).build();
-        let pep_input = PEPRelation::try_from(input)?;
+        let pup_input = PUPRelation::try_from(input)?;
 
-        let (dp_relation, private_query) = pep_input
+        let (dp_relation, private_query) = pup_input
             .differentially_private_sums(
                 named_sums
                     .iter() // Convert &str to String
@@ -220,8 +221,8 @@ impl Reduce {
         epsilon: f64,
         delta: f64,
     ) -> Result<DPRelation> {
-        let pep_input = PEPRelation::try_from(self.input().clone())?;
-        pep_input.differentially_private_aggregates(
+        let pup_input = PUPRelation::try_from(self.input().clone())?;
+        pup_input.differentially_private_aggregates(
             self.named_aggregates()
                 .into_iter()
                 .map(|(n, agg)| (n, agg.clone()))
@@ -242,7 +243,7 @@ mod tests {
         data_type::Variant,
         display::Dot,
         io::{postgresql, Database},
-        protection::{Protection, Strategy},
+        privacy_unit_tracking::{PrivacyUnitTracking, Strategy},
         sql::parse,
         Relation,
     };
@@ -282,8 +283,8 @@ mod tests {
             .clone();
         let (epsilon, delta) = (1., 1e-3);
 
-        // protect the inputs
-        let protection = Protection::from((
+        // privacy tracking of the inputs
+        let privacy_unit_tracking = PrivacyUnitTracking::from((
             &relations,
             vec![
                 (
@@ -295,18 +296,20 @@ mod tests {
             ],
             Strategy::Hard,
         ));
-        let pep_table = protection.table(&table.try_into().unwrap()).unwrap();
+        let pup_table = privacy_unit_tracking
+            .table(&table.try_into().unwrap())
+            .unwrap();
 
         let reduce = Reduce::new(
             "my_reduce".to_string(),
             vec![("sum_price".to_string(), AggregateColumn::sum("price"))],
             vec![],
-            pep_table.deref().clone().into(),
+            pup_table.deref().clone().into(),
         );
         let relation = Relation::from(reduce.clone());
         relation.display_dot().unwrap();
 
-        let dp_relation = PEPRelation::try_from(reduce.input().clone())
+        let dp_relation = PUPRelation::try_from(reduce.input().clone())
             .unwrap()
             .differentially_private_sums(vec!["price"], vec![], epsilon, delta)
             .unwrap();
@@ -330,8 +333,8 @@ mod tests {
             .clone();
         let (epsilon, delta) = (1., 1e-3);
 
-        // protect the inputs
-        let protection = Protection::from((
+        // privacy tracking of the inputs
+        let privacy_unit_tracking = PrivacyUnitTracking::from((
             &relations,
             vec![
                 (
@@ -343,18 +346,20 @@ mod tests {
             ],
             Strategy::Hard,
         ));
-        let pep_table = protection.table(&table.try_into().unwrap()).unwrap();
+        let pup_table = privacy_unit_tracking
+            .table(&table.try_into().unwrap())
+            .unwrap();
 
         let reduce = Reduce::new(
             "my_reduce".to_string(),
             vec![("my_sum_price".to_string(), AggregateColumn::sum("price"))],
             vec![Expr::col("item")],
-            pep_table.deref().clone().into(),
+            pup_table.deref().clone().into(),
         );
         let relation = Relation::from(reduce.clone());
         relation.display_dot().unwrap();
 
-        let dp_relation = PEPRelation::try_from(reduce.input().clone())
+        let dp_relation = PUPRelation::try_from(reduce.input().clone())
             .unwrap()
             .differentially_private_sums(vec!["price"], vec!["item"], epsilon, delta)
             .unwrap();
@@ -382,8 +387,8 @@ mod tests {
             .clone();
         let (epsilon, delta) = (1., 1e-3);
 
-        // protect the inputs
-        let protection = Protection::from((
+        // privacy tracking of the inputs
+        let privacy_unit_tracking = PrivacyUnitTracking::from((
             &relations,
             vec![
                 (
@@ -395,7 +400,9 @@ mod tests {
             ],
             Strategy::Hard,
         ));
-        let pep_table = protection.table(&table.try_into().unwrap()).unwrap();
+        let pup_table = privacy_unit_tracking
+            .table(&table.try_into().unwrap())
+            .unwrap();
         let reduce = Reduce::new(
             "my_reduce".to_string(),
             vec![
@@ -404,7 +411,7 @@ mod tests {
                 ("avg_price".to_string(), AggregateColumn::mean("price")),
             ],
             vec![],
-            pep_table.deref().clone().into(),
+            pup_table.deref().clone().into(),
         );
         let relation = Relation::from(reduce.clone());
         relation.display_dot().unwrap();
@@ -439,8 +446,8 @@ mod tests {
             .clone();
         let (epsilon, delta) = (1., 1e-3);
 
-        // protect the inputs
-        let protection = Protection::from((
+        // privacy tracking of the inputs
+        let privacy_unit_tracking = PrivacyUnitTracking::from((
             &relations,
             vec![
                 (
@@ -452,7 +459,9 @@ mod tests {
             ],
             Strategy::Hard,
         ));
-        let pep_table = protection.table(&table.try_into().unwrap()).unwrap();
+        let pup_table = privacy_unit_tracking
+            .table(&table.try_into().unwrap())
+            .unwrap();
         let reduce = Reduce::new(
             "my_reduce".to_string(),
             vec![
@@ -461,7 +470,7 @@ mod tests {
                 ("avg_price".to_string(), AggregateColumn::mean("price")),
             ],
             vec![expr!(item)],
-            pep_table.deref().clone().into(),
+            pup_table.deref().clone().into(),
         );
         let relation = Relation::from(reduce.clone());
         relation.display_dot().unwrap();
@@ -497,7 +506,7 @@ mod tests {
                 ("avg_price".to_string(), AggregateColumn::mean("price")),
             ],
             vec![Expr::col("item")],
-            pep_table.deref().clone().into(),
+            pup_table.deref().clone().into(),
         );
         let relation = Relation::from(reduce.clone());
         relation.display_dot().unwrap();