+++ src/expr/sql.rs
@@ -555,8 +555,6 @@ mod tests {
         println!("ast::expr = {gen_expr}");
         assert_eq!(gen_expr, parse_expr("substr(a, 0, 5)").unwrap());
     }
-
-
     #[test]
     fn test_cast() {
         let str_expr = "cast(a as varchar)";